refactor: setup extensions, engines and sidecar
