+++ src-tauri/.gitignore
@@ -2,3 +2,5 @@
 # will have compiled files and executables
 /target/
 /gen/schemas
+binaries
+!binaries/download.sh
