+++ src-tauri/src/core/mod.rs
@@ -1,2 +1,3 @@
 pub mod cmd;
 pub mod fs;
+pub mod setup;
